Extend AePageObjects::Node with #element for declaring elements dynamically
This allows you to create elements on the fly. It's useful for things like
modals which cannot be interacted with until they are triggered by some other
action (e.g. a button click). Instead of declaring the modal as a "constant"
element of the Document, you would define a method that corresponds to the
action that opens the modal, and this method can yield a dynamically created
ElementProxy for interaction with the modal.

So, instead of:

  class ShowPage
    element :share_modal, is: ShareModal
  end

(which makes instances of ShowPage *always* expose access to a share_modal
element via show_page.share_modal, even when the modal has not been triggered)

You'd have:

  class ShowPage
    def share_image(&block)
      node.click_button('Share')
      share_modal = document.element(locator: '#share_modal', is: ShareModal)
      share_modal.wait_until_visible
      block.call(share_modal)
      share_modal.wait_until_hidden
    end
  end

which only exposes the share_image method on instances of ShowPage. Thus, users
can only access a share_modal element in the block this method yields to, which
is the only time the modal can *actually* be interacted with.

Huzzah! Error prevention via a well-designed page object whose interface
matches the real-world user interface!